+++ src/01.Domain/Core/Shop.Domain.Core/OrderAgg/Entities/OrderItem.cs
@@ -14,10 +14,10 @@ namespace Shop.Domain.Core.OrderAgg.Entities
         public int ProductId { get; set; }
         public Product Product { get; set; }
 
-        public double UnitPrice { get; set; }
+        public decimal UnitPrice { get; set; }
 
         public int Quantity { get; set; }
 
-        public double TotalPrice => UnitPrice * Quantity;
+        public decimal TotalPrice => UnitPrice * Quantity;
     }
 }
